+++ src/jvmMain/kotlin/creals/impl/AssumedIntReal.kt
@@ -1,8 +1,7 @@
 package garden.ephemeral.calculator.creals.impl
 
+import garden.ephemeral.calculator.bigint.BigInt
 import garden.ephemeral.calculator.creals.Real
-import garden.ephemeral.calculator.creals.util.scale
-import java.math.BigInteger
 
 /**
  * Representation of a number that may not have been completely
@@ -10,7 +9,7 @@ import java.math.BigInteger
  * evaluate beyond the decimal point.
  */
 internal class AssumedIntReal(val value: Real) : Real() {
-    override fun approximate(precision: Int): BigInteger {
+    override fun approximate(precision: Int): BigInt {
         return if (precision >= 0) {
             value.getApproximation(precision)
         } else {
