+++ src/jvmMain/kotlin/creals/impl/NegationReal.kt
@@ -1,13 +1,13 @@
 package garden.ephemeral.calculator.creals.impl
 
+import garden.ephemeral.calculator.bigint.BigInt
 import garden.ephemeral.calculator.creals.Real
-import java.math.BigInteger
 
 /**
  * Representation of the negation of a constructive real.
  */
 internal class NegationReal(private val op: Real) : Real() {
-    override fun approximate(precision: Int): BigInteger {
+    override fun approximate(precision: Int): BigInt {
         return -op.getApproximation(precision)
     }
 }
