+++ src/jvmMain/kotlin/creals/impl/ShiftedReal.kt
@@ -1,13 +1,13 @@
 package garden.ephemeral.calculator.creals.impl
 
+import garden.ephemeral.calculator.bigint.BigInt
 import garden.ephemeral.calculator.creals.Real
-import java.math.BigInteger
 
 /**
  * Representation of a constructive real multiplied by `2**n`.
  */
 internal class ShiftedReal(private val op: Real, private val count: Int) : Real() {
-    override fun approximate(precision: Int): BigInteger {
+    override fun approximate(precision: Int): BigInt {
         return op.getApproximation(precision - count)
     }
 }
